requirements.txt
# As we approach deployment, we should probably move to different requirements files for local vs. production
# See Chapt 5.5 in Two Scoops of Django
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.8
django-bootstrap3==4.11.0
django-cors-headers==1.1.0
django-crispy-forms==1.4.0
django-toolbelt==0.0.1
django-twitter-bootstrap==3.3.0
djangorestframework==3.1.2
elasticsearch>=1.0.0,<2.0.0
future==0.15.2
google-api-python-client==1.4.0
gunicorn==19.3.0
httplib2==0.9.1
oauth2client==1.4.7
oauthlib==0.7.2
nameparser==0.3.11
psycopg2==2.6
pyasn1==0.1.7
pyasn1-modules==0.0.5
py3dns==3.1.0
pygeoip==0.3.2
geopy==1.11.0
PyJWT==1.3.0
python-openid==2.2.5
python-social-auth==0.2.10
requests==2.6.2
requests-oauthlib==0.5.0
rsa==3.1.4
selenium==2.45.0
simplejson==3.6.5
six==1.9.0
static3==0.5.1
tweepy==3.5.0
Unipath==1.1
uritemplate==0.6
usaddress==0.5.5
validate-email==1.3
# python-votesmart=0.3.3  # Using local version instead since 0.3.3 not Py 3 compatible
wikipedia==1.4.0
xmltodict==0.9.2
